Clean up ANSI warnings
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1406 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack.d/apply.c
@@ -208,7 +208,7 @@ in_ice_box(obj) register struct obj *obj; {
 		pline("You cannot refrigerate something you are wearing.");
 		return(0);
 	}
-	if(obj->owt + current_ice_box->owt > 70) {
+	if(obj->owt + current_ice_box->owt > (unsigned)70) {
 		pline("It won't fit.");
 		return(1);	/* be careful! */
 	}
